requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
lightgbm
imbalanced-learn
shap
optuna
fastapi
uvicorn
pydantic
joblib
streamlit
plotly
jupyter
python-dotenv
groq
sqlalchemy
passlib
python-jose
bcrypt